Update web UI with pastel vortex animation and matching monitor style
- Add slowly spinning ASCII vortex background with pastel colors
- 3 wide spiral arms with smooth rotation
- Monitor restyled with lavender/purple pastel theme
- Improved screen sizing (800x500)
- Auto-focus canvas for keyboard input

+++ include/mm/pmm.h
@@ -14,7 +14,7 @@
 #define PAGE_ALIGN_UP(x) (((x) + PAGE_SIZE - 1) & ~(PAGE_SIZE - 1))
 
 // Initialize physical memory manager
-void pmm_init(uint64_t memory_size);
+void pmm_init(uint32_t memory_size);
 
 // Allocate a single physical page
 void* pmm_alloc_page(void);
